feat(bolao): tabs Tabela e Ranking no BolaoEmptyState
Sub-nav agora tem 4 entradas: [Palpites] [Especiais] [Tabela] [Ranking].

- Palpites (default): conteudo atual (Hero "Faca seus N palpites" + Convidar)
- Especiais: continua abrindo SpecialPredictionsModal (Campeao + finalistas)
- Tabela: 12 grupos da Copa 2026 com J/V/E/D/Pts via CopaGruposView
- Ranking: BolaoRankingTable inline (empty state ja tem CTA "Convidar amigos"
  que volta pra aba Palpites onde mora o card de convite)

Refactor: CopaGruposView extraido de CopaGruposModal — agora a view standalone
serve tanto pro modal (BolaoHome) quanto pra tab inline. CopaGruposModal vira
um wrapper Dialog fino em volta da view.

- BolaoEmptyState recebe nova prop `ranking: BolaoRankingEntry[] | undefined`
  passada pelo BolaoDetail (que ja consulta useBolaoRanking).
- Sub-nav overflow-x-auto pra caber em mobile sem quebrar.
